backend_requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
tensorflow>=2.9.0
Pillow>=9.0.0
pandas>=1.3.0
numpy>=1.19.0
python-multipart>=0.0.5
